requirements.txt
strudel.scraper
pandas
xlrd
openpyxl
PyGithub